move hybrid search to its own component
